Refactor: useUserControl 확장 위해 분리
