+++ src/styles/globals.css
@@ -54,3 +54,7 @@ input {
 #scroll-box::-webkit-scrollbar-thumb {
   @apply rounded-full bg-gray-600;
 }
+
+.userControl-item {
+  @apply w-full p-2 cursor-pointer hover:bg-gray-700;
+}
